@@ -26,7 +26,7 @@
   "version": "2.8.0-alpha.1",
   "dependencies": {
     "@fortawesome/fontawesome-free": "^5.12.1",
-    "alignment.js": "0.2.1",
+    "alignment.js": "0.3.0",
     "bootstrap": "4.x",
     "chi-squared": "^1.1.0",
     "circos": "^2.1.0",
@@ -49,11 +49,11 @@
     "popper.js": "^1.14.1",
     "pretty-data": "^0.40.0",
     "prop-types": "^15.5.10",
-    "react": "16",
+    "react": "16.12.0",
     "react-copy-to-clipboard": "^5.0.0",
-    "react-dom": "16",
+    "react-dom": "16.12.0",
     "react-json-view": "^1.14.0",
-    "react-phylotree": "0.1.0",
+    "react-phylotree": "0.4.1",
     "react-range": "^1.5.3",
     "react-router-dom": "^4.2.2",
     "react-router-hash-link": "^1.1.1",
